+++ experimental/mocks/service/dataquality/mock_data_quality_interface.go
@@ -7,7 +7,6 @@ import (
 
 	listing "github.com/databricks/databricks-sdk-go/listing"
 	dataquality "github.com/databricks/databricks-sdk-go/service/dataquality"
-
 	mock "github.com/stretchr/testify/mock"
 )